fix(frontend): separate useAppSection hook to fix Fast Refresh
